requirements.txt
Django==2.2
django-tinymce==2.8.0
pkg-resources==0.0.0
psycopg2-binary==2.8.1
pytz==2018.9
sqlparse==0.3.0
